move vod items to vod.ts
