+++ web/composables/utils.ts
@@ -4,15 +4,6 @@ export const BASE_URL = import.meta.env.PROD
     ? "https://api.mercuryland.pp.ua"
     : "http://127.0.0.1:8080";
 
-export interface VodItem {
-    id: number;
-    date: string;
-    link: string;
-    title: string;
-    tags: string[];
-    duration: string;
-}
-
 export interface PenItem {
     id: number;
     date: string;
@@ -37,27 +28,6 @@ export function ofId<T extends { id: number }>(data: T[], id: number): T {
     return null;
 }
 
-// parse HH:MM:SS format into seconds
-export function parseHMS(hms: string): number {
-    let [s = 0, m = 0, h = 0] = hms.split(":").reverse();
-    return +h * 3600 + +m * 60 + +s;
-}
-
-export function formatHMS(seconds: number) {
-    seconds = Math.abs(seconds);
-    let s = seconds % 60;
-    let m = Math.floor(seconds / 60) % 60;
-    let h = Math.floor(seconds / 3600);
-
-    return (
-        h.toString().padStart(2, "0") +
-        ":" +
-        m.toString().padStart(2, "0") +
-        ":" +
-        s.toString().padStart(2, "0")
-    );
-}
-
 export function formatDate(date: Date): string {
     let year = date.getFullYear();
     let month = date.getMonth() + 1;
@@ -76,29 +46,9 @@ export function parseDate(text): Date {
     return new Date(year, month - 1, day);
 }
 
-export function remainingX(el: Element) {
-    return (
-        document.documentElement.clientWidth -
-        el.getBoundingClientRect().left +
-        window.screenX
-    );
-}
-
-export function remainingY(el) {
-    return (
-        document.documentElement.clientHeight -
-        el.getBoundingClientRect().top +
-        window.scrollY
-    );
-}
-
 export function backToTop() {
     window.scrollTo({ top: 0, behavior: "smooth" });
 }
-
-// export function interleave<T>(arr: T[], x: T): T[] {
-//     return arr.flatMap((e) => [e, x]).slice(0, -1);
-// }
 
 export async function copyToClipboard(text: string) {
     try {
